Move `controller.provide_pid_for_xml_uri()` para `sources.harvesting.provide_pid_for_xml_uri()` E elimina importações sem uso
